move analyze compression patterns
